pyproject.toml
[project]
name = "Water-Info-Acquirer"
version = "1.2.0"
description = "公開水文データから水位・流量・雨量データを指定期間ごとに取得する。"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "pyperclip>=1.11.0",
    "pyyaml>=6.0.3",
    "matplotlib>=3.10.0",
    "requests>=2.32.5",
    "xlsxwriter>=3.2.9",
]

[dependency-groups]
dev = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.0",
    "pyinstaller>=6.16.0",
    "pytest>=9.0.1",
    "ruff>=0.15.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]
exclude = ["src/jma_rainfall_pipeline"]

[tool.ruff.lint]
select = ["F", "E", "W"]
ignore = ["E501"]
